app/src/main/java/com/dicoding/picodiploma/moviecatalogue/ui/detail/movie/DetailMovieViewModel.kt
package com.dicoding.picodiploma.moviecatalogue.ui.detail.movie

import androidx.lifecycle.LiveData
import androidx.lifecycle.ViewModel
import com.dicoding.picodiploma.moviecatalogue.MovieRepository
import com.dicoding.picodiploma.moviecatalogue.data.Model

class DetailMovieViewModel(private val repository: MovieRepository) : ViewModel() {
    private lateinit var movieId: String

    fun setSelectedMovie(movieId: String) {
        this.movieId = movieId
    }

    fun getMovie(movieId: String): LiveData<Model> = repository.getMovieWithModules(movieId)
}